Mapping Extensions v0.25.0
ID: MappingExtensions  
Author: StackDoubleFlow, rxzz0, & rcelyte  
Mod Loader: Scotland2  
Game Version: 1.40.4_5283  

--------------------------------------------------

This adds a host of new things you can do with your maps as a mapper, and allows you to play said maps as a player. An update of the port of the PC original mod by Kyle 1413. Previously maintained by zoller27osu.